refactor: Structural refactor: move npa.workbench → npa.solutions.workbench with platform docs
Task-ID: task-cli-46b88c02-ba92-43a3-be9a-b9e88cc75647
Dry-Run: false
