+++ npa/src/npa/cli/workbench/lancedb/backfill.py
@@ -4,13 +4,16 @@ from __future__ import annotations
 
 import typer
 
-from npa.workbench.lancedb.backfill import (
+from npa.solutions.workbench.lancedb.backfill import (
     DEFAULT_BATCH_SIZE,
     DEFAULT_DHASH_HAMMING_THRESHOLD,
     BackfillError,
     backfill_column,
 )
-from npa.workbench.lancedb.bdd100k_import import DEFAULT_LANCE_URI, DEFAULT_TABLE
+from npa.solutions.workbench.lancedb.bdd100k_import import (
+    DEFAULT_LANCE_URI,
+    DEFAULT_TABLE,
+)
 
 from .helpers import (
     DEFAULT_TOKEN_ENV,
@@ -33,17 +36,31 @@ def backfill_cmd(
         "--output-path",
         help="LanceDB URI containing the table; input and output are the same table.",
     ),
-    batch_size: int = typer.Option(DEFAULT_BATCH_SIZE, "--batch-size", help="Rows per CPU batch."),
-    force: bool = typer.Option(False, "--force", help="Recompute rows that already have non-null values."),
+    batch_size: int = typer.Option(
+        DEFAULT_BATCH_SIZE, "--batch-size", help="Rows per CPU batch."
+    ),
+    force: bool = typer.Option(
+        False, "--force", help="Recompute rows that already have non-null values."
+    ),
     dhash_hamming_threshold: int = typer.Option(
         DEFAULT_DHASH_HAMMING_THRESHOLD,
         "--dhash-hamming-threshold",
         help="Near-duplicate threshold for udf=is_duplicate.",
     ),
-    service: bool = typer.Option(False, "--service", help="Call a deployed LanceDB workbench endpoint."),
-    endpoint: str = typer.Option("", "--endpoint", help="LanceDB wrapper endpoint for --service."),
-    token_env: str = typer.Option(DEFAULT_TOKEN_ENV, "--token-env", help="Environment variable containing wrapper token."),
-    output: OutputFormat = typer.Option(OutputFormat.json, "--output", help="Output format."),
+    service: bool = typer.Option(
+        False, "--service", help="Call a deployed LanceDB workbench endpoint."
+    ),
+    endpoint: str = typer.Option(
+        "", "--endpoint", help="LanceDB wrapper endpoint for --service."
+    ),
+    token_env: str = typer.Option(
+        DEFAULT_TOKEN_ENV,
+        "--token-env",
+        help="Environment variable containing wrapper token.",
+    ),
+    output: OutputFormat = typer.Option(
+        OutputFormat.json, "--output", help="Output format."
+    ),
 ) -> None:
     """Backfill one BDD100K UDF column through local mode or a deployed service."""
     payload = {
