+++ npa/src/npa/cli/workbench/lancedb/views.py
@@ -4,8 +4,17 @@ from __future__ import annotations
 
 import typer
 
-from npa.workbench.lancedb.bdd100k_import import DEFAULT_LANCE_URI, DEFAULT_TABLE
-from npa.workbench.lancedb.views import DEFAULT_QUERY_LIMIT, MVError, create_mv, query_table, refresh_mv
+from npa.solutions.workbench.lancedb.bdd100k_import import (
+    DEFAULT_LANCE_URI,
+    DEFAULT_TABLE,
+)
+from npa.solutions.workbench.lancedb.views import (
+    DEFAULT_QUERY_LIMIT,
+    MVError,
+    create_mv,
+    query_table,
+    refresh_mv,
+)
 
 from .helpers import (
     DEFAULT_TOKEN_ENV,
@@ -20,8 +29,12 @@ from .helpers import (
 
 def create_mv_cmd(
     name: str = typer.Option(..., "--name", help="Materialized view table name."),
-    source_table: str = typer.Option(DEFAULT_TABLE, "--source", "--source-table", help="Source LanceDB table name."),
-    filter_sql: str = typer.Option(..., "--filter", "--filter-sql", help="SQL WHERE clause for the source table."),
+    source_table: str = typer.Option(
+        DEFAULT_TABLE, "--source", "--source-table", help="Source LanceDB table name."
+    ),
+    filter_sql: str = typer.Option(
+        ..., "--filter", "--filter-sql", help="SQL WHERE clause for the source table."
+    ),
     lance_uri: str = typer.Option(
         DEFAULT_LANCE_URI,
         "--lance-uri",
@@ -29,11 +42,23 @@ def create_mv_cmd(
         "--output-path",
         help="LanceDB URI containing the source and view tables.",
     ),
-    force: bool = typer.Option(False, "--force", help="Recompute and overwrite an existing registered view."),
-    service: bool = typer.Option(False, "--service", help="Call a deployed LanceDB workbench endpoint."),
-    endpoint: str = typer.Option("", "--endpoint", help="LanceDB wrapper endpoint for --service."),
-    token_env: str = typer.Option(DEFAULT_TOKEN_ENV, "--token-env", help="Environment variable containing wrapper token."),
-    output: OutputFormat = typer.Option(OutputFormat.json, "--output", help="Output format."),
+    force: bool = typer.Option(
+        False, "--force", help="Recompute and overwrite an existing registered view."
+    ),
+    service: bool = typer.Option(
+        False, "--service", help="Call a deployed LanceDB workbench endpoint."
+    ),
+    endpoint: str = typer.Option(
+        "", "--endpoint", help="LanceDB wrapper endpoint for --service."
+    ),
+    token_env: str = typer.Option(
+        DEFAULT_TOKEN_ENV,
+        "--token-env",
+        help="Environment variable containing wrapper token.",
+    ),
+    output: OutputFormat = typer.Option(
+        OutputFormat.json, "--output", help="Output format."
+    ),
 ) -> None:
     """Create a filtered materialized view through local mode or a deployed service."""
     payload = {
@@ -75,10 +100,20 @@ def refresh_mv_cmd(
         "--output-path",
         help="LanceDB URI containing the registry and view table.",
     ),
-    service: bool = typer.Option(False, "--service", help="Call a deployed LanceDB workbench endpoint."),
-    endpoint: str = typer.Option("", "--endpoint", help="LanceDB wrapper endpoint for --service."),
-    token_env: str = typer.Option(DEFAULT_TOKEN_ENV, "--token-env", help="Environment variable containing wrapper token."),
-    output: OutputFormat = typer.Option(OutputFormat.json, "--output", help="Output format."),
+    service: bool = typer.Option(
+        False, "--service", help="Call a deployed LanceDB workbench endpoint."
+    ),
+    endpoint: str = typer.Option(
+        "", "--endpoint", help="LanceDB wrapper endpoint for --service."
+    ),
+    token_env: str = typer.Option(
+        DEFAULT_TOKEN_ENV,
+        "--token-env",
+        help="Environment variable containing wrapper token.",
+    ),
+    output: OutputFormat = typer.Option(
+        OutputFormat.json, "--output", help="Output format."
+    ),
 ) -> None:
     """Refresh a registered materialized view through local mode or a deployed service."""
     payload = {"name": name, "lance_uri": lance_uri}
@@ -107,19 +142,35 @@ def refresh_mv_cmd(
 
 def query_table_cmd(
     table: str = typer.Option(..., "--table", help="LanceDB table name to query."),
-    filter_sql: str = typer.Option("", "--filter", "--filter-sql", help="Optional SQL WHERE clause."),
-    select: list[str] = typer.Option([], "--select", help="Column to return; repeatable."),
-    limit: int = typer.Option(DEFAULT_QUERY_LIMIT, "--limit", help="Maximum rows to return."),
+    filter_sql: str = typer.Option(
+        "", "--filter", "--filter-sql", help="Optional SQL WHERE clause."
+    ),
+    select: list[str] = typer.Option(
+        [], "--select", help="Column to return; repeatable."
+    ),
+    limit: int = typer.Option(
+        DEFAULT_QUERY_LIMIT, "--limit", help="Maximum rows to return."
+    ),
     lance_uri: str = typer.Option(
         DEFAULT_LANCE_URI,
         "--lance-uri",
         "--input-path",
         help="LanceDB URI containing the table.",
     ),
-    service: bool = typer.Option(False, "--service", help="Call a deployed LanceDB workbench endpoint."),
-    endpoint: str = typer.Option("", "--endpoint", help="LanceDB wrapper endpoint for --service."),
-    token_env: str = typer.Option(DEFAULT_TOKEN_ENV, "--token-env", help="Environment variable containing wrapper token."),
-    output: OutputFormat = typer.Option(OutputFormat.json, "--output", help="Output format."),
+    service: bool = typer.Option(
+        False, "--service", help="Call a deployed LanceDB workbench endpoint."
+    ),
+    endpoint: str = typer.Option(
+        "", "--endpoint", help="LanceDB wrapper endpoint for --service."
+    ),
+    token_env: str = typer.Option(
+        DEFAULT_TOKEN_ENV,
+        "--token-env",
+        help="Environment variable containing wrapper token.",
+    ),
+    output: OutputFormat = typer.Option(
+        OutputFormat.json, "--output", help="Output format."
+    ),
 ) -> None:
     """Run a bounded SQL-filtered LanceDB table query."""
     payload = {
