+++ npa/src/npa/sdk/workbench/lancedb/__init__.py
@@ -2,19 +2,19 @@
 
 from __future__ import annotations
 
-from npa.workbench.lancedb import (
-    BDD100KImportError,
+from npa.solutions.workbench.lancedb import (
+    BDD100KImportError as BDD100KImportError,
     BDD100KImportResult,
-    BDD100KServiceError,
-    BDD100KValidationError,
-    BackfillError,
+    BDD100KServiceError as BDD100KServiceError,
+    BDD100KValidationError as BDD100KValidationError,
+    BackfillError as BackfillError,
     BackfillResult,
-    BackfillServiceError,
-    BackfillValidationError,
-    MVError,
+    BackfillServiceError as BackfillServiceError,
+    BackfillValidationError as BackfillValidationError,
+    MVError as MVError,
     MVResult,
-    MVServiceError,
-    MVValidationError,
+    MVServiceError as MVServiceError,
+    MVValidationError as MVValidationError,
     QueryResult,
     backfill,
     create_bdd100k_failure_mode_views,
