+++ npa/src/npa/solutions/__init__.py
@@ -1,0 +1,3 @@
+"""Solution namespaces for Nebius Physical AI."""
+
+from __future__ import annotations
