+++ npa/src/npa/solutions/workbench/__init__.py
@@ -1,0 +1,16 @@
+"""npa.solutions.workbench - deployable Physical AI workbench tools."""
+
+from __future__ import annotations
+
+from . import cosmos, detection_training, fiftyone, genesis, groot, isaac_lab, lancedb, lerobot
+
+__all__ = [
+    "cosmos",
+    "detection_training",
+    "fiftyone",
+    "genesis",
+    "groot",
+    "isaac_lab",
+    "lancedb",
+    "lerobot",
+]
