+++ npa/src/npa/solutions/workbench/detection_training/__init__.py
@@ -1,0 +1,14 @@
+"""npa.solutions.workbench.detection_training - Faster R-CNN detection training tool."""
+
+from __future__ import annotations
+
+from .schemas import EvalRequest, EvalResponse, StatusResponse, TrainRequest, TrainResponse
+
+__all__ = [
+    "EvalRequest",
+    "EvalResponse",
+    "StatusResponse",
+    "TrainRequest",
+    "TrainResponse",
+]
+
