+++ npa/src/npa/workbench/isaac_lab/__init__.py
@@ -1,29 +1,7 @@
-"""npa.workbench.isaac_lab - Isaac Lab workbench commands."""
+"""Deprecated compatibility shim for npa.solutions.workbench.isaac_lab."""
 
 from __future__ import annotations
 
-from npa._sdk import make_cli_wrapper
+from npa.workbench._compat import install_shim
 
-list = make_cli_wrapper("npa.cli.isaac_lab", "list_cmd", "List Isaac Lab workbenches.")
-deploy = make_cli_wrapper(
-    "npa.cli.isaac_lab", "deploy_cmd", "Deploy an Isaac Lab workbench."
-)
-status = make_cli_wrapper("npa.cli.isaac_lab", "status_cmd", "Show Isaac Lab status.")
-system_info = make_cli_wrapper(
-    "npa.cli.isaac_lab", "system_info_cmd", "Show Isaac Lab system information."
-)
-train = make_cli_wrapper("npa.cli.isaac_lab", "train_cmd", "Train in Isaac Lab.")
-eval = make_cli_wrapper("npa.cli.isaac_lab", "eval_cmd", "Evaluate in Isaac Lab.")
-export_lerobot = make_cli_wrapper(
-    "npa.cli.isaac_lab", "export_lerobot_cmd", "Export Isaac Lab rollouts as LeRobot."
-)
-
-__all__ = [
-    "list",
-    "deploy",
-    "status",
-    "system_info",
-    "train",
-    "eval",
-    "export_lerobot",
-]
+install_shim(__name__, "npa.solutions.workbench.isaac_lab", globals())
